css changes

+++ myplanttracker/item/models.py
@@ -1,6 +1,7 @@
 from django.contrib.auth.models import User 
 from django.db import models
 
+
 # Create your models here.
 
 class Category(models.Model):
@@ -17,7 +18,7 @@ class Item(models.Model):
     category = models.ForeignKey(Category, related_name ='items', on_delete=models.CASCADE)
     image = models.ImageField(upload_to='item_images', blank=True, null=True)
     name = models.CharField(max_length=255)
-    price = models.FloatField()
+    price= models.DecimalField(max_digits=10, decimal_places=2)
     location = models.CharField(max_length=255, null=True,blank=True)
     description = models.TextField(blank=True, null=True)
     is_sold = models.BooleanField(default=False)
